Okay, bnf labors... go fix elidables
I got "ui klama" working. But now, "ti broda" doesn't match. (It matches ti, but not fullly.)
So, why doesn't the broda match?
Because bridi-tail doesn't match. BUT, {ti broda vau} DOES match. So, it's probably has to do with elidable terminators not being dealt with properly! So, I'm about to try to fix the elidable terminators.